feat(update car status): update car advert status [Finishes #166726256]
