+++ .babelrc
@@ -1,5 +1,3 @@
 {
-    "plugins": [
-      ["@babel/plugin-proposal-class-properties", { "loose": true }]
-    ]
+    "presets": ["@babel/preset-env"]
 }
