+++ .gitignore
@@ -1,1 +1,3 @@
 node_modules/
+.nyc_output/
+coverage/
