+++ API/controllers/cars/carsController.js
@@ -18,7 +18,7 @@ class CarsController{
         }  
         
         pool.connect((err, client, done) => { 
-            const query = "INSERT INTO cars(owner, state, status, price, manufacturer, model, body_type) VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING * ";
+            const query = "INSERT INTO cars(owner, state, status, price, manufacturer, model, body_type) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING * ";
              const values = [data.owner, data.state, data.status, data.price, data.manufacturer, data.model, data.body_type];
 
              client.query(query, values, (error, result) => {
@@ -30,28 +30,29 @@ class CarsController{
                 return res.status(200).send({ status: 200, message: 'Car Ad successfully saved', data: result.rows[0] })
               });
         })
-
     }
 
-    static postCarAd(req, res) 
+    static updateAdvertStatus(req, res)
     {
-        const {
-            body: {
-                owner,  
-                created_on,
-                state,
-                status,
-                price,   
-                manufacturer,
-                model,
-                body_type
-            }            
-        } = req;
+        const id = parseInt(req.params.id)
+        const data = {
+            status: req.body.status
+        }
 
-        const result = cars.newCar(owner, created_on, state, status, price, manufacturer, model, body_type);
+        pool.connect((err, client, done) => 
+        {
+            const query = "UPDATE cars SET status=$1 WHERE id=$2";
+            const values = [data.status, id];
 
-        return res.status(200).json({ status: 200, data: result});
-  
+            client.query(query, values, (error, result) => {
+                done();
+                if(error)
+                {
+                    res.status(400).json({ status: 400, message: error });
+                }
+                return res.status(200).send({ status: 200, message: 'Car status successfully updated', data: result.rows[0] })
+            });
+        })
     }
 
 }
