+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "nodemon ./api/server/index",
     "build": "rm -rf ./build && babel -d ./build ./API -s",
-    "test": "mocha ./api/test/",
+    "test": " nyc --reporter=html --reporter=text mocha --require @babel/register './api/test/tests.js' --exit",
     "create": "node ./api/services/db createTables",
     "generate-lcov": "nyc report --reporter=text-lcov > lcov.info",
     "coveralls-coverage": "coveralls < lcov.info",
@@ -17,11 +17,14 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@babel/cli": "^7.0.0",
+    "@babel/cli": "^7.4.4",
     "@babel/core": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/node": "^7.4.5",
+    "@babel/plugin-transform-runtime": "^7.4.4",
     "@babel/preset-env": "^7.4.5",
     "@babel/register": "^7.4.4",
+    "@hapi/joi": "^15.0.3",
+    "babel-eslint": "^10.0.1",
     "babel-polyfill": "^6.26.0",
     "body-parser": "^1.19.0",
     "chai": "^4.2.0",
